Test 12

+++ Chatopia/src/com/zaidock/chat/utills/GameObject.java
@@ -9,7 +9,7 @@ public abstract class GameObject {
 
 	protected float x, y;
 	protected ID id;
-	protected float velX, velY;
+	protected int velX, velY;
 
 	public boolean showingHitBoxes = false;
